+++ src/Positionable.tsx
@@ -51,14 +51,26 @@ export interface PositionableProps {
     /** Render Prop alternative to using `children` */
     render?: RenderCallback;
 
-    /** Snap drag and resize to pixels of this interval. */
-    snapTo?: number;
-
     /** Should resizing be enabled? */
     resizable?: boolean;
 
     /** Should rotation be enabled? */
     rotatable?: boolean;
+
+    /** Snap drag and resize to pixels of this interval. */
+    snapTo?: number;
+
+    /**
+     * Snap X drag and resize to pixels of this interval (Overwrite snapTo in X dimension).
+     * set snapX as `0` to restrict moving/resizing to vertical
+     */
+    snapX?: number;
+
+    /**
+     * Snap Y drag and resize to pixels of this interval (Overwrite snapTo in Y dimension).
+     * set snapY as `0` to restrict moving/resizing to horizontal
+     */
+    snapY?: number;
 }
 
 type RenderCallback = (args: RenderCallbackArgs) => JSX.Element;
@@ -179,6 +191,8 @@ export class Positionable extends React.Component<
             resizable,
             rotatable,
             snapTo,
+            snapX,
+            snapY,
         } = this.props;
         const { left, width } = this.state;
         const group = this.props.group || randomString();
@@ -201,6 +215,8 @@ export class Positionable extends React.Component<
                     this.refHandlers.dnd.current ||
                     this.refHandlers.container.current,
                 snapTo,
+                snapX,
+                snapY,
             })
                 .pipe(takeUntil(this.destroy$))
                 .subscribe();
@@ -245,6 +261,8 @@ export class Positionable extends React.Component<
                     left: config.left,
                     shouldConvertToPercent: width.includes('%'),
                     snapTo,
+                    snapX,
+                    snapY,
                 })
                     .pipe(takeUntil(this.destroy$))
                     .subscribe(newPosition => this.setState(newPosition));