+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- `snapX` and `snapY` in combination with `snapTo`. This supports restriction on dragging and resizing horizontally (snapY === 0), or vertically (snapX === 0)
+
 ## [0.4.1] - 2019-02-11
 ### Changed
 - When using `snapTo` with grouped elements, the element that is being